+++ src/abstract-authenticated-command.js
@@ -1,14 +1,31 @@
 const { Command } = require('@oclif/command');
-const chalk = require('chalk');
-const logger = require('./services/logger');
-const authenticator = require('./services/authenticator');
+const context = require('./context');
 
 class AbstractAuthenticatedCommand extends Command {
+  constructor(...args) {
+    super(...args);
+    /** @type {import('./context/init').Context} */
+    const { logger, authenticator, chalk } = context.inject();
+
+    /** @protected @readonly */
+    this.logger = logger;
+
+    /** @protected @readonly */
+    this.authenticator = authenticator;
+
+    /** @protected @readonly */
+    this.chalk = chalk;
+
+    ['logger', 'authenticator', 'chalk'].forEach((name) => {
+      if (!this[name]) throw new Error(`Missing dependency ${name}`);
+    });
+  }
+
   async run() {
-    if (!authenticator.getAuthToken()) {
-      logger.info('Login required.');
-      await authenticator.tryLogin({});
-      if (!authenticator.getAuthToken()) this.exit(10);
+    if (!this.authenticator.getAuthToken()) {
+      this.logger.info('Login required.');
+      await this.authenticator.tryLogin({});
+      if (!this.authenticator.getAuthToken()) this.exit(10);
     }
 
     try {
@@ -16,12 +33,12 @@ class AbstractAuthenticatedCommand extends Command {
     } catch (error) {
       // NOTICE: Due to ip-whitelist, 404 will never be thrown for a project
       if (error.status === 403) {
-        logger.error('You do not have the right to execute this action on this project');
+        this.logger.error('You do not have the right to execute this action on this project');
         return this.exit(2);
       }
 
       if (error.status === 401) {
-        await authenticator.logout();
+        await this.authenticator.logout();
         return this.displayLoginMessageAndQuit();
       }
 
@@ -34,7 +51,7 @@ class AbstractAuthenticatedCommand extends Command {
   }
 
   displayLoginMessageAndQuit() {
-    logger.error(`Please use ${chalk.bold('forest login')} to sign in to your Forest account.`);
+    this.logger.error(`Please use ${this.chalk.bold('forest login')} to sign in to your Forest account.`);
     this.exit(10);
   }
 }