+++ src/commands/deploy.js
@@ -6,10 +6,25 @@ const ProjectManager = require('../services/project-manager');
 const { handleBranchError } = require('../services/branch-manager');
 const withCurrentProject = require('../services/with-current-project');
 
-const { inquirer, config: envConfig } = context.inject();
-
 /** Deploy layout changes of an environment to production. */
 class DeployCommand extends AbstractAuthenticatedCommand {
+  constructor(...args) {
+    super(...args);
+
+    /** @type {import('../context/init').Context} */
+    const { inquirer, config } = context.inject();
+
+    /** @private @readonly */
+    this.inquirer = inquirer;
+
+    /** @private @readonly */
+    this.envConfig = config;
+
+    ['inquirer', 'envConfig'].forEach((name) => {
+      if (!this[name]) throw new Error(`Missing dependency ${name}`);
+    });
+  }
+
   /**
    * Get selected environment; prompt for environment when none is provided.
    * @param {Object} config - Actual command config (including command parameters).
@@ -17,7 +32,7 @@ class DeployCommand extends AbstractAuthenticatedCommand {
    * @throws Will throw an error when there is no production environment.
    * @return {Object} The environment found.
    */
-  static async getEnvironment(config) {
+  async getEnvironment(config) {
     const environments = await new EnvironmentManager(config).listEnvironments();
     const productionExists = environments.find((environment) => environment.type === 'production');
 
@@ -25,7 +40,7 @@ class DeployCommand extends AbstractAuthenticatedCommand {
     if (!productionExists) throw new Error('❌ You need a production environment to run this command.');
 
     const environmentName = config.ENVIRONMENT_NAME
-      || await DeployCommand.selectEnvironment(environments);
+      || await this.selectEnvironment(environments);
     return environments.find((environment) => environment.name === environmentName);
   }
 
@@ -34,8 +49,8 @@ class DeployCommand extends AbstractAuthenticatedCommand {
    * @param {Array} environments List of environments (from backend).
    * @see getEnvironment
    */
-  static async selectEnvironment(environments) {
-    const response = await inquirer.prompt([{
+  async selectEnvironment(environments) {
+    const response = await this.inquirer.prompt([{
       name: 'environment',
       message: 'Select the environment containing the layout changes you want to deploy to production',
       type: 'list',
@@ -55,7 +70,7 @@ class DeployCommand extends AbstractAuthenticatedCommand {
     const envSecret = process.env.FOREST_ENV_SECRET;
     const parsed = this.parse(DeployCommand);
     const commandOptions = { ...parsed.flags, ...parsed.args, envSecret };
-    const config = await withCurrentProject({ ...envConfig, ...commandOptions });
+    const config = await withCurrentProject({ ...this.envConfig, ...commandOptions });
 
     if (!config.envSecret) {
       const environment = await new ProjectManager(config)
@@ -71,8 +86,8 @@ class DeployCommand extends AbstractAuthenticatedCommand {
    * @param {Object} environment - The environment containing the layout changes to deploy.
    * @returns {Boolean} Return true if user has confirmed.
    */
-  static async confirm(environment) {
-    const response = await inquirer
+  async confirm(environment) {
+    const response = await this.inquirer
       .prompt([{
         type: 'confirm',
         name: 'confirm',
@@ -88,11 +103,11 @@ class DeployCommand extends AbstractAuthenticatedCommand {
   async runIfAuthenticated() {
     try {
       const config = await this.getConfig();
-      const environment = await DeployCommand.getEnvironment(config);
+      const environment = await this.getEnvironment(config);
 
       if (environment === undefined) throw new Error('Environment not found.');
       if (environment.type === 'production') throw new Error('❌ You cannot deploy production onto itself.');
-      if (!config.force && !(await DeployCommand.confirm(environment))) return null;
+      if (!config.force && !(await this.confirm(environment))) return null;
 
       await new EnvironmentManager(config).deploy(environment);