+++ src/services/api.js
@@ -1,13 +1,16 @@
-const agent = require('superagent');
-const pkg = require('../../package.json');
-
 const HEADER_CONTENT_TYPE = 'Content-Type';
 const HEADER_CONTENT_TYPE_JSON = 'application/json';
 const HEADER_FOREST_ORIGIN = 'forest-origin';
 const HEADER_USER_AGENT = 'User-Agent';
 
-function API() {
-  this.endpoint = () => process.env.FOREST_URL || 'https://api.forestadmin.com';
+/**
+ * @class
+ * @param {import('../context/init').Context} context
+ */
+function Api({
+  pkg, env, superagent: agent, applicationTokenSerializer, applicationTokenDeserializer,
+}) {
+  this.endpoint = () => env.FOREST_URL || 'https://api.forestadmin.com';
   this.userAgent = `forest-cli@${pkg.version}`;
   const headers = {
     [HEADER_FOREST_ORIGIN]: 'forest-cli',
@@ -15,18 +18,36 @@ function API() {
     [HEADER_USER_AGENT]: this.userAgent,
   };
 
-  this.isGoogleAccount = async (email) => agent
-    .get(`${this.endpoint()}/api/users/google/${email}`)
-    .set(headers)
-    .send()
-    .then((response) => response.body.data.isGoogleAccount)
-    .catch(() => false);
-
   this.login = async (email, password) => agent
     .post(`${this.endpoint()}/api/sessions`)
     .set(headers)
     .send({ email, password })
     .then((response) => response.body.token);
+
+  /**
+   * @param {import('../serializers/application-token').InputApplicationToken} applicationToken
+   * @param {string} sessionToken
+   * @returns {Promise<import('../deserializers/application-token').ApplicationToken>}
+   */
+  this.createApplicationToken = async (applicationToken, sessionToken) => agent
+    .post(`${this.endpoint()}/api/application-tokens`)
+    .set(headers)
+    .set('Authorization', `Bearer ${sessionToken}`)
+    .send(applicationTokenSerializer.serialize(applicationToken))
+    .then((response) => applicationTokenDeserializer.deserialize(response.body));
+
+
+  /**
+   * @param {string} sessionToken
+   * @returns {Promise<import('../deserializers/application-token').ApplicationToken>}
+   */
+  this.deleteApplicationToken = async (sessionToken) => agent
+    .delete(`${this.endpoint()}/api/application-tokens`)
+    .set(HEADER_FOREST_ORIGIN, 'forest-cli')
+    .set(HEADER_CONTENT_TYPE, HEADER_CONTENT_TYPE_JSON)
+    .set(HEADER_USER_AGENT, this.userAgent)
+    .set('Authorization', `Bearer ${sessionToken}`)
+    .send();
 }
 
-module.exports = new API();
+module.exports = Api;