+++ src/services/project-manager.js
@@ -1,11 +1,13 @@
 const _ = require('lodash');
 const P = require('bluebird');
 const agent = require('superagent-promise')(require('superagent'), P);
-const authenticator = require('./authenticator');
+const context = require('../context');
 const ProjectSerializer = require('../serializers/project');
 const ProjectDeserializer = require('../deserializers/project');
 const EnvironmentDeserializer = require('../deserializers/environment');
 const { serverHost } = require('../config');
+
+const { authenticator } = context.inject();
 
 function ProjectManager(config) {
   function deserialize(response) {