+++ test/commands/test-cli-std.js
@@ -1,7 +1,7 @@
 const mockStdin = require('mock-stdin');
 const { stdout, stderr } = require('stdout-stderr');
 
-const TIME_TO_LOAD_TEST = 1000;
+const TIME_TO_LOAD_TEST = 2000;
 const TIME_TO_REACH_NEXT_PROMPT = 100;
 
 module.exports = {
@@ -12,16 +12,27 @@ module.exports = {
     stderr.print = print;
     const stdin = mockStdin.stdin();
     if (outputs.length) stdout.start();
-    if (errorOutputs.length) stderr.start();
+    stderr.start();
     return stdin;
   },
+  /**
+   * @param {string[]} inputs
+   * @param {import('mock-stdin').MockSTDIN} stdin
+   */
   planifyInputs: (inputs, stdin) => {
     for (let i = 0; i < inputs.length; i += 1) {
       // WARNING: Smelly code..., could break if the test pre-loading time increases.
       setTimeout(() => stdin.send(`${inputs[i]}\n`), TIME_TO_LOAD_TEST + i * TIME_TO_REACH_NEXT_PROMPT);
     }
   },
-  assertOutputs: (outputs, errorOutputs) => {
+  assertOutputs: (outputs, errorOutputs, { assertNoStdError }) => {
+    stdout.stop();
+    stderr.stop();
+
+    if (!errorOutputs.length && stderr.output && assertNoStdError) {
+      expect(stderr.output.trim()).toBe('');
+    }
+
     for (let i = 0; i < outputs.length; i += 1) {
       const isString = typeof outputs[i] === 'string' || outputs[i] instanceof String;
       if (isString) {